[feat] : Add translations

+++ adapters/web/src/main/kotlin/com/esgi/nova/adapters/web/endpoints/games/GamesLocation.kt
@@ -5,5 +5,5 @@ import io.ktor.locations.*
 
 @KtorExperimentalLocationsAPI
 @Location("/games")
-data class GamesLocation(override val page: Int, override val size: Int) : IPagination {
+data class GamesLocation(override val page: Long, override val size: Long) : IPagination {
 }
