+++ ports/src/main/kotlin/com/esgi/nova/ports/provided/IPagination.kt
@@ -1,6 +1,6 @@
 package com.esgi.nova.ports.provided
 
 interface IPagination {
-    val page: Int
-    val size: Int
+    val page: Long
+    val size: Long
 }
